requirements.txt
# Video Factory - Python Dependencies

# Core
python-dotenv>=1.0.0

# Video Processing
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# AI Services
google-generativeai>=0.3.0
requests>=2.31.0

# Audio Processing
pydub>=0.25.1

# Utilities
Pillow>=10.0.0
tqdm>=4.66.0
